changed structure of the project, including adding app file with separated views, database and html
