+++ .flake8
@@ -1,3 +1,3 @@
 [flake8]
 max-line-length = 88
-extend-ignore = E203, E704
+extend-ignore = E203, E704, F401, E402
