+++ .gitignore
@@ -1,4 +1,4 @@
 .DS_Store
 .idea/
 .env
-instance
+__pycache__
