+++ app/__init__.py
@@ -1,0 +1,6 @@
+from flask import Flask
+
+app = Flask(__name__)
+
+from app import database
+from app import views
