+++ app/database.py
@@ -1,0 +1,30 @@
+from app import app
+from flask_sqlalchemy import SQLAlchemy
+from flask_wtf import FlaskForm
+from wtforms import StringField
+from wtforms.validators import DataRequired
+import os
+
+app.config["SECRET_KEY"] = os.getenv("SECRET_KEY")
+app.config["SQLALCHEMY_DATABASE_URI"] = os.getenv("SQLALCHEMY_DATABASE_URI")
+db = SQLAlchemy()
+db.init_app(app)
+
+
+class Movie(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    title = db.Column(db.String, unique=True)
+    year = db.Column(db.Integer)
+    description = db.Column(db.String)
+    rating = db.Column(db.Float)
+    ranking = db.Column(db.Integer)
+    review = db.Column(db.String)
+    img_url = db.Column(db.String)
+
+
+class MyForm(FlaskForm):
+    title = StringField("Movie Title", validators=[DataRequired()])
+
+
+with app.app_context():
+    db.create_all()
