+++ app/templates/edit.html
@@ -4,16 +4,16 @@
 
 {% block content %}
 <div class="content">
-  <h1 class="heading">Movie Title</h1>
+  <h1 class="heading">{{ title }}</h1>
     <p class="description">Edit Movie Rating</p>
 
     <form method="POST">
       <div class="mb-3">
-        <label for="rating" class="form-label">Your New Rating: </label>
+        <label for="rating" class="form-label">Your Rating: </label>
         <input type="text" class="form-control" id="rating" name="rating">
       </div>
       <div class="mb-3">
-        <label for="review" class="form-label">Your New Review: </label>
+        <label for="review" class="form-label">Your Review: </label>
         <input type="text" class="form-control" id="review" name="review">
       </div>
       <button type="submit" class="btn btn-primary">Submit</button>
